:books: Update finetuner documentation
