+++ modelci/finetuner/tl_algorithm.py
@@ -36,9 +36,6 @@ Reference:
     https://github.com/PyTorchLightning/pytorch-lightning/blob/master/pl_examples/domain_templates/computer_vision_fine_tuning.py
 """
 
-import argparse
-from pathlib import Path
-from tempfile import TemporaryDirectory
 from typing import Generator, Optional, Callable
 
 import pytorch_lightning as pl
@@ -47,9 +44,6 @@ from torch import optim
 from torch.nn import Module
 from torch.optim.lr_scheduler import StepLR
 from torch.optim.optimizer import Optimizer
-
-from modelci.finetuner.torch_data_module import PyTorchDataModule
-from modelci.finetuner.trainer import PyTorchTrainer
 
 BN_TYPES = (torch.nn.BatchNorm1d, torch.nn.BatchNorm2d, torch.nn.BatchNorm3d)
 
@@ -147,8 +141,8 @@ def _unfreeze_and_add_param_group(
 #  --- Pytorch-lightning module ---
 
 
-class TransferLearningModel(pl.LightningModule):
-    """Transfer Learning with pre-trained ResNet50."""
+class FineTuneModule(pl.LightningModule):
+    """Transfer Learning with pre-trained model"""
 
     def __init__(
             self,
@@ -229,90 +223,3 @@ class TransferLearningModel(pl.LightningModule):
         scheduler = StepLR(optimizer, step_size=self.step_size, gamma=self.lr_scheduler_gamma)
 
         return [optimizer], [scheduler]
-
-
-def main(args: argparse.Namespace) -> None:
-    """Train the model.
-
-    Args:
-        args: Model hyper-parameters
-
-    Note:
-        For the sake of the example, the images dataset will be downloaded
-        to a temporary directory.
-    """
-
-    with TemporaryDirectory(dir=args.root_data_path):
-        # TODO: Done in editor (L247-251)
-        net = torch.hub.load('pytorch/vision:v0.6.0', args.backbone, pretrained=True)
-        freeze(module=net, train_bn=True)
-
-        num_ftrs = net.fc.in_features
-        net.fc = torch.nn.Linear(num_ftrs, 10)
-
-        model = TransferLearningModel(**vars(args), net=net, loss=torch.nn.CrossEntropyLoss())
-        data_module = PyTorchDataModule('CIFAR10', batch_size=args.batch_size, data_dir=args.root_data_path)
-
-        trainer = PyTorchTrainer(
-            model=model,
-            data_loader_kwargs={'datamodule': data_module},
-            trainer_kwargs={
-                'weights_summary': None,
-                'progress_bar_refresh_rate': 1,
-                'num_sanity_val_steps': 0,
-                'gpus': args.gpus,
-                'min_epochs': args.nb_epochs,
-                'max_epochs': args.nb_epochs,
-            }
-        )
-        trainer.start()
-        trainer.join()
-
-
-def get_args() -> argparse.Namespace:
-    parser = argparse.ArgumentParser(add_help=False)
-    parser.add_argument(
-        "--root-data-path",
-        metavar="DIR",
-        type=str,
-        default=Path.cwd().as_posix(),
-        help="Root directory where to download the data",
-        dest="root_data_path",
-    )
-    parser.add_argument(
-        "--backbone",
-        default="resnet50",
-        type=str,
-        metavar="BK",
-        help="Name (as in ``torchvision.models``) of the feature extractor",
-    )
-    parser.add_argument(
-        "--epochs", default=15, type=int, metavar="N", help="total number of epochs", dest="nb_epochs"
-    )
-    parser.add_argument("--batch-size", default=128, type=int, metavar="B", help="batch size", dest="batch_size")
-    parser.add_argument("--gpus", type=int, default=1, help="number of gpus to use")
-    parser.add_argument(
-        "--lr", "--learning-rate", default=1e-2, type=float, metavar="LR", help="initial learning rate", dest="lr"
-    )
-    parser.add_argument(
-        "--lr-scheduler-gamma",
-        default=1e-1,
-        type=float,
-        metavar="LRG",
-        help="Factor by which the learning rate is reduced at each milestone",
-    )
-    parser.add_argument(
-        "--step-size",
-        default=7,
-        type=int,
-        metavar="SS",
-        help="Step size used in step scheduler",
-    )
-    parser.add_argument(
-        "--num-workers", default=6, type=int, metavar="W", help="number of CPU workers", dest="num_workers"
-    )
-    return parser.parse_args()
-
-
-if __name__ == "__main__":
-    main(get_args())
